+++ docs/assets/UluFormItemsInline.stories-BtAB7Lha.js
@@ -1,0 +1,43 @@
+import{_ as o}from"./UluFormItemsInline-Cjb1qupP.js";import{_ as l}from"./UluFormItem-DdCBjxD0.js";import{_ as r}from"./UluFormCheckbox-op3pDPa3.js";import"./iframe-kPZklDD1.js";import"./preload-helper-BJwshlQW.js";import"./UluFormLabel-LIulJAsg.js";import"./UluFormRequiredChar--PWDqaZ9.js";import"./props-DEaRQ31f.js";const U={component:o,tags:["autodocs"]},e={render:m=>({components:{UluFormItemsInline:o,UluFormItem:l,UluFormCheckbox:r},setup(){return{args:m}},template:`
+      <div class="form-theme">
+        <UluFormItemsInline>
+          <UluFormItem layout="checkbox" label="Option 1">
+            <UluFormCheckbox :modelValue="true" />
+          </UluFormItem>
+          <UluFormItem layout="checkbox" label="Option 2">
+            <UluFormCheckbox :modelValue="false" />
+          </UluFormItem>
+          <UluFormItem layout="checkbox" label="Option 3">
+            <UluFormCheckbox :modelValue="true" />
+          </UluFormItem>
+        </UluFormItemsInline>
+      </div>
+    `})};e.parameters={...e.parameters,docs:{...e.parameters?.docs,source:{originalSource:`{
+  render: args => ({
+    components: {
+      UluFormItemsInline,
+      UluFormItem,
+      UluFormCheckbox
+    },
+    setup() {
+      return {
+        args
+      };
+    },
+    template: \`
+      <div class="form-theme">
+        <UluFormItemsInline>
+          <UluFormItem layout="checkbox" label="Option 1">
+            <UluFormCheckbox :modelValue="true" />
+          </UluFormItem>
+          <UluFormItem layout="checkbox" label="Option 2">
+            <UluFormCheckbox :modelValue="false" />
+          </UluFormItem>
+          <UluFormItem layout="checkbox" label="Option 3">
+            <UluFormCheckbox :modelValue="true" />
+          </UluFormItem>
+        </UluFormItemsInline>
+      </div>
+    \`
+  })
+}`,...e.parameters?.docs?.source}}};const I=["Default"];export{e as Default,I as __namedExportsOrder,U as default};