+++ docs/assets/props-DEaRQ31f.js
@@ -1,0 +1,1 @@
+function f(e){return e.every(c=>typeof c=="object")}function o(e,c,t){c.forEach(r=>{e[r]!==void 0&&t(r)})}export{o as c,f as i};